Change the return type of Block's size-related methods to long
Currently the getSizeInBytes/getRegionSizeInBytes/getRetainedSizeInBytes
methods all return integers. To implement these methods Block implementations
usually sum up the size of their internal states and cast that sum to integer
using different cast functions, which are different in different Block implementations,
and occasionally incorrect (using Ints.saturatedCast()).

+++ presto-main/src/main/java/com/facebook/presto/operator/GroupByIdBlock.java
@@ -57,7 +57,7 @@ public class GroupByIdBlock
     }
 
     @Override
-    public int getRegionSizeInBytes(int positionOffset, int length)
+    public long getRegionSizeInBytes(int positionOffset, int length)
     {
         return block.getRegionSizeInBytes(positionOffset, length);
     }
@@ -171,13 +171,13 @@ public class GroupByIdBlock
     }
 
     @Override
-    public int getSizeInBytes()
+    public long getSizeInBytes()
     {
         return block.getSizeInBytes();
     }
 
     @Override
-    public int getRetainedSizeInBytes()
+    public long getRetainedSizeInBytes()
     {
         return INSTANCE_SIZE + block.getRetainedSizeInBytes();
     }
