+++ presto-spi/src/main/java/com/facebook/presto/spi/block/AbstractSingleArrayBlock.java
@@ -168,7 +168,7 @@ public abstract class AbstractSingleArrayBlock
     }
 
     @Override
-    public int getRegionSizeInBytes(int position, int length)
+    public long getRegionSizeInBytes(int position, int length)
     {
         throw new UnsupportedOperationException();
     }
